Divide application service into it and DTO
