+++ frontend/components/comment/CommentList.vue
@@ -40,7 +40,7 @@
 import Vue, { PropType } from 'vue'
 import VueFilterDateFormat from '@vuejs-community/vue-filter-date-format'
 import VueFilterDateParse from '@vuejs-community/vue-filter-date-parse'
-import { CommentReadDTO } from '@/services/application/comment.service'
+import { CommentReadDTO } from '~/services/application/comment/commentData'
 Vue.use(VueFilterDateFormat)
 Vue.use(VueFilterDateParse)
