+++ frontend/plugins/services.ts
@@ -8,25 +8,25 @@ import { FromApiCommentItemListRepository } from '@/repositories/comment/api'
 import { FromApiStatisticsRepository } from '@/repositories/statistics/api'
 import { FromApiDocumentItemListRepository } from '@/repositories/document/api'
 import { LocalStorageOptionRepository} from '@/repositories/option/api'
-import { LabelApplicationService } from '@/services/application/label.service'
-import { MemberApplicationService } from '@/services/application/member.service'
-import { UserApplicationService } from '@/services/application/user.service'
-import { RoleApplicationService } from '@/services/application/role.service'
-import { ProjectApplicationService } from '@/services/application/project.service'
-import { CommentApplicationService } from '@/services/application/comment.service'
-import { StatisticsApplicationService } from '@/services/application/statistics.service'
-import { DocumentApplicationService } from '@/services/application/document.service'
-import { OptionApplicationService } from '@/services/application/option.service'
 import { FromApiSequenceLabelingRepository } from '@/repositories/tasks/sequenceLabeling/api'
-import { SequenceLabelingApplicationService } from '@/services/application/tasks/sequenceLabelingService'
 import { FromApiSeq2seqRepository } from '@/repositories/tasks/seq2seq/api'
-import { Seq2seqApplicationService } from '@/services/application/tasks/seq2seqService'
-import { ConfigApplicationService } from '@/services/application/config.service'
 import { FromApiConfigItemListRepository } from '@/repositories/config/api'
-import { TemplateApplicationService } from '@/services/application/template.service'
 import { FromApiTemplateRepository } from '@/repositories/template/api'
+import { LabelApplicationService } from '~/services/application/label/labelApplicationService'
+import { MemberApplicationService } from '~/services/application/member/memberApplicationService'
+import { UserApplicationService } from '~/services/application/user/userApplicationService'
+import { RoleApplicationService } from '~/services/application/role/roleApplicationService'
+import { ProjectApplicationService } from '~/services/application/project/projectApplicationService'
+import { CommentApplicationService } from '~/services/application/comment/commentApplicationService'
+import { StatisticsApplicationService } from '~/services/application/statistics/statisticsApplicationService'
+import { DocumentApplicationService } from '~/services/application/document/documentApplicationService'
+import { OptionApplicationService } from '~/services/application/option/optionApplicationService'
+import { SequenceLabelingApplicationService } from '~/services/application/tasks/sequenceLabeling/sequenceLabelingApplicationService'
+import { Seq2seqApplicationService } from '~/services/application/tasks/seq2seq/seq2seqApplicationService'
+import { ConfigApplicationService } from '~/services/application/autoLabeling/configApplicationService'
+import { TemplateApplicationService } from '~/services/application/autoLabeling/templateApplicationService'
 import { FromApiTextClassificationRepository } from '~/repositories/tasks/textClassification/api'
-import { TextClassificationApplicationService } from '~/services/application/tasks/textClassificationService'
+import { TextClassificationApplicationService } from '~/services/application/tasks/textClassification/textClassificationApplicationService'
 
 
 export interface Services {
