+++ frontend/services/application/autoLabeling/configApplicationService.ts
