+++ frontend/services/application/autoLabeling/templateApplicationService.ts
