+++ frontend/services/application/comment/commentApplicationService.ts
@@ -1,23 +1,6 @@
 import { CommentItem } from '@/models/comment'
 import { CommentItemListRepository } from '@/repositories/comment/interface'
-
-export class CommentReadDTO {
-  id: number
-  user: number
-  username: string
-  documentText: string
-  text: string
-  createdAt: string
-
-  constructor(item: CommentItem) {
-    this.id = item.id
-    this.user = item.user
-    this.username = item.username
-    this.documentText = item.documentText
-    this.text = item.text
-    this.createdAt = item.createdAt
-  }
-}
+import { CommentReadDTO } from './commentData'
 
 export class CommentApplicationService {
   constructor(
