+++ frontend/services/application/comment/commentData.ts
@@ -1,0 +1,20 @@
+import { CommentItem } from '@/models/comment'
+
+
+export class CommentReadDTO {
+  id: number;
+  user: number;
+  username: string;
+  documentText: string;
+  text: string;
+  createdAt: string;
+
+  constructor(item: CommentItem) {
+    this.id = item.id;
+    this.user = item.user;
+    this.username = item.username;
+    this.documentText = item.documentText;
+    this.text = item.text;
+    this.createdAt = item.createdAt;
+  }
+}
