+++ frontend/services/application/label/labelApplicationService.ts
@@ -1,23 +1,7 @@
 import { LabelItem } from '@/models/label'
 import { LabelItemListRepository } from '@/repositories/label/interface'
+import { LabelDTO } from './labelData'
 
-export class LabelDTO {
-  id: number
-  text: string
-  prefixKey: string | null
-  suffixKey: string | null
-  backgroundColor: string
-  textColor: string
-
-  constructor(item: LabelItem) {
-    this.id = item.id
-    this.text = item.text
-    this.prefixKey = item.prefixKey
-    this.suffixKey = item.suffixKey
-    this.backgroundColor = item.backgroundColor
-    this.textColor = '#ffffff'
-  }
-}
 
 export class LabelApplicationService {
   constructor(
