+++ frontend/services/application/member/memberApplicationService.ts
@@ -1,21 +1,6 @@
 import { MemberItem } from '@/models/member'
 import { MemberItemListRepository } from '@/repositories/member/interface'
-
-export class MemberDTO {
-  id: number
-  user: number
-  role: number
-  username: string
-  rolename: string
-
-  constructor(item: MemberItem) {
-    this.id = item.id
-    this.user = item.user
-    this.role = item.role
-    this.username = item.username
-    this.rolename = item.rolename
-  }
-}
+import { MemberDTO } from './memberData'
 
 export class MemberApplicationService {
   constructor(
