+++ frontend/services/application/statistics/statisticsApplicationService.ts
@@ -1,17 +1,5 @@
-import { Statistics } from '@/models/statistics'
 import { StatisticsRepository } from '@/repositories/statistics/interface'
-
-export class StatisticsDTO {
-  label: object
-  user: object
-  progress: object
-
-  constructor(item: Statistics, labelText: string, userText: string, progressLabels: string[]) {
-    this.label = item.labelStats(labelText)
-    this.user = item.userStats(userText)
-    this.progress = item.progress(progressLabels)
-  }
-}
+import { StatisticsDTO } from './statisticsData'
 
 export class StatisticsApplicationService {
   constructor(
