+++ frontend/services/application/tasks/seq2seq/seq2seqData.ts
@@ -1,0 +1,14 @@
+import { Seq2seqLabel } from '@/models/tasks/seq2seq'
+
+
+export class Seq2seqDTO {
+  id: number;
+  text: string;
+  user: number;
+
+  constructor(item: Seq2seqLabel) {
+    this.id = item.id;
+    this.text = item.text;
+    this.user = item.user;
+  }
+}
