+++ frontend/services/application/tasks/textClassification/textClassificationData.ts
@@ -1,0 +1,14 @@
+import { TextClassificationItem } from '~/models/tasks/textClassification'
+
+
+export class TextClassificationDTO {
+  id: number;
+  label: number;
+  user: number;
+
+  constructor(item: TextClassificationItem) {
+    this.id = item.id;
+    this.label = item.label;
+    this.user = item.user;
+  }
+}
